64 bit int/float

+++ src/rust_function_ctx.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub enum RustValue {
-    Float(f32),
-    Integer(i32),
+    Float(f64),
+    Integer(i64),
     Str(String),
     Bool(bool),
     StringVector(Vec<String>),
@@ -10,8 +10,8 @@ pub enum RustValue {
 pub trait RustFunctionCtx {
     fn get_parameter(&self, idx: usize) -> Option<RustValue>;
 
-    fn get_parameter_float(&self, idx: usize) -> Option<f32>;
-    fn get_parameter_integer(&self, idx: usize) -> Option<i32>;
+    fn get_parameter_float(&self, idx: usize) -> Option<f64>;
+    fn get_parameter_integer(&self, idx: usize) -> Option<i64>;
     fn get_parameter_bool(&self, idx: usize) -> Option<bool>;
     fn get_parameter_string(&self, idx: usize) -> Option<String>;
